+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.0.5] - 2015-09-25
 - Use host from collectd for Top Info notification
+- Add mem_total for physical memory in kb
 
 ## [0.0.4] - 2015-09-20
 - Top on amazon linux and better error handling